net: lib: azure_iot_hub: Update API version
Update the API version for the MQTT connection to Azure IoT Hub.
This fixes the issue where JSON messages containing arrays are not
accepted when updating the device twin. This is due to arrays not
being supported in the validation logic used by the IoT Hub in
previous API versions.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>