samples: nrf9160: coap_client: update response handling
If a response with no payload is returned, the software reboots.
The crash happens in the snprintf function.
The function coap_packet_get_payload returns NULL
if there is no payload.
The snprintf function tries then to copy from a NULL pointer.

This happens if the get request is not set up correctly,
for example a wrong URI_PATH is defined.
The server then responds with an error code and an empty payload.

Signed-off-by: Matthias Ludwig <ludw@zhaw.ch>