samples: nrf9160: lwm2m_client: NB-IoT update early
Increase the value from default 6s to 30s. Time in seconds before the
registration timeout, when the LWM2M Registration Update is sent by
the engine. In networks with large round trip times (like NB-IoT), it
might be needed to set this value higher, in order to allow the
response to arrive before timeout.

Signed-off-by: Veijo Pesonen <veijo.pesonen@nordicsemi.no>