Bluetooth: Mesh: Use fast cadence when value exceedes delta
Section 4.3.1.2.4.3 of E15886 and E15551 makes Sensor server use Fast
Cadence Period Divisor to publish sensor values if the change exceeds
delta defined by Status Trigger Delta Up/Down and Type states:

The message shall be published more frequently when the value
of the measured quantity is lower than the previously published
value decremented by the value of the Status Trigger Delta Down
state (see Section 4.1.3.3) or when it is higher than the previously
published value incremented by the value of the Status Trigger Delta
Up state (see Section 4.1.3.4). When the message is to be published
more frequently, the Publish Period value is divided by the value of
the Fast Cadence Period Divisor state as described in Section 3.1.3.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>