Bluetooth: Mesh: Enable the controller only after the first boot
Enforcing the controller to be enabled at the start overwrites the
behavior of the Generic OnPowerUp state, thus the correct state of
the light ctrl is not recovered after reboot. This also breaks
MMDL/SR/LLC/BV-11-C.

This change enables the controller in the sample only after the first
boot so that it is possible to interact with the controller without a
light ctrl client. After reboot, the controller's state will depend on
the Generic OnPowerUp state.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>