mpsl: Add support for 1wire coex
Added KConfig for the 1wire coex protocol to easily enable this feature.
Added device-tree support for the 1-wire coex feature. This allows the
configuration of all 1wire coex parameters through the device-tree.
1-wire coex will be marked as an experimental feature.

doc: Updated documentation for 1-wire coex

Split documentation for the bluetooth coexistence feature into a part
for the 3wire and one for the 1wire coexistence protocol.

Signed-off-by: Jan Müller <jan.mueller@nordicsemi.no>
Signed-off-by: Krishna Badam <krishna.badam@nordicsemi.no>
Signed-off-by: Gopi Gunda <gopi.gunda@nordicsemi.no>