caf: sensor_manager: Correct initialization behaviour
During the initialization of sensors, if any fails to initialize,
the module reports error and disallows to sample remaining sensors.

This behaviour is rather an inconvinience; plus it was not aligned
with module documentation.

Fix this by allowing sensor_manager to continue sampling remaining
sensors in case some of them fail to initialize.

Jira: NCSDK-11645

Signed-off-by: Emil Obalski <emil.obalski@nordicsemi.no>