doc: add nRF9160 DK getting started info
Add the information from Infocenter on getting started
with the nRF9160 DK next to NCS' Working with the nRF9160 DK
section.

Removed section on modem firmware from Working with
the nRF9160 DK section since the topic is covered in
the Getting Started section.

Moved section on provisioning nRF Cloud certificates from
the getting started to Working with nRF9160 DK.

Also added related entries to the glossary, and some links.

Signed-off-by: Wille Backman <wille.backman@nordicsemi.no>