spm: kconfig: use SYSTICK as system clock driver instead of RTC1
RTC1 is configured as non-secure by default, and in some configurations
a fault would be triggered if RTC1 was used as the system clock driver
if it was configured as non-secure.

Fix this by using SYSTICK as the system clock driver instead.

Also add some Kconfig to avoid users using RTC1 as system clock driver
for SPM when RTC1 is configured as non-secure.

Ref: NCSDK-12230

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>