treewide: re-work DK buttons and leds library
Rework to build gpio_dt_spec structures for the devicetree child nodes
of /leds and /buttons to capture the board-specific state, and use it
in the library.

We have buttons at /buttons and leds at /leds on all our DK boards.
Use this fact to avoid hard-coding individual buttons and leds by
devicetree alias, instead just iterating over every available button
and LED on the board at preprocessor time.

We also have active low or high flags set up for each individual
button or LED in each board's devicetree. Use this to eliminate the
global "invert" Kconfig option by getting the information on a
per-button or per-LED basis from the devicetree.

Any bad behavior which may be introduced by this change is a bug in
the board's devicetree, which should be fixed there instead of being
worked around (e.g. global Kconfig options like
CONFIG_DK_LIBRARY_INVERT_LEDS).

This avoids boilerplate in both the library and the applications, and
makes the library more customizable in case individual buttons or LEDs
have strange behavior.

Keep in-tree users up to date.

Ref: NCSDK-13380

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>