lib: fota_download: fixed fota update fail after link timeout
fota_download would in some cases fail to restart the download from
the right offset after a link connection timeout. This would result
in an invalid image, and the whole fota update would have to be
restarted.

Reason for failure:
When an amount of data that is not a multiple of
CONFIG_FOTA_DOWNLOAD_MCUBOOT_FLASH_BUF_SZ is written with
stream_flash_buffered_write() this leaves data in the stream
flash buffer between calls.
Stream_flash_bytes_written() only reports number of bytes written to
flash (in multiples of CONFIG_FOTA_DOWNLOAD_MCUBOOT_FLASH_BUF_SZ) and
does not account for bytes written to the internal stream flash buffer.

When fota_download uses dfu_target_mcuboot_offset_get() to get
download offset for download_client, it can result in a series of
bytes being duplicated and consecutive bytes being offseted in flash.
Resulting in a wrong firmware image.

Fix:
In dfu_target_mcuboot, keep a counter of bytes in the stream
flash buffer in order to correct the offset returned by
dfu_target_mcuboot_offset_get().

NCSDK-12930

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>
Signed-off-by: Francesco Domenico Servidio <francesco.servidio@nordicsemi.no>