lib: at_monitor: introduce AT_MONITOR_ISR
Introduce AT_MONITOR_ISR macro to define monitors that receive
notifications directly, that is, in the same ISR they are sent
from by libmodem.

These monitors avoid the situation where there is no memory to copy
the notification on the library heap. This is useful for applications
that must ensure to ack certain notifications (SMS-related) and
need a guarantee that they can always be received.

Importantly, the application may not call other nrf_modem_at
APIs in the ISR context where notifications are received.
The application must ensure to defer acking any notifications
received in the ISR outside of the ISR.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>