tfm: Fix ARRAY_SIZE redefined by util.h in out-of-tree board support
Fix ARRAY_SIZE redefined by util.h in out-of-tree board support.
RTE_Device is an out-fo-tree TF-M header that is included by the
in-tree TF-M board support and includes zephyr headers in turn.
This causes the ARRAY_SIZE macro defined in TF-M array.h to be
redefined by util.h

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>