net: lib: azure_iot_hub: Add warning about risk of reconnections
Add a warning in the Azure IoT Hub library about the known
connectivity issue with Azure IoT Hub when the device is not
reachable.
The issue is not limited to the Asset Tracker v2, so the
Known issues note is moved to the "Other issues" section to reflect
that. Suggested workaround added to the note.

Closes CIA-511.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>