bootloader: bl_crypto: Increase hash context size and fix test
After Oberon crypto library was updated to 3.0.9 the context size
of the  sha256 functions was increased. This commit updates the size
used in `bl_crypto` so that `ocrypt_sha256_ctx` can fit inside the
external crypto context and can be used from another image.

This commit also updates the tests to earlier report failure and fix
a faulty check for `NULL` in `data` when the size is 0.

Ref. NCSDK-13498

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>