lib: nrf_modem: Remove auto stopping of traces from modem_trace module
It is currently not possible to call nrf_modem_at_printf from interrupt
context. This happened when the trace session timer expired leading to
an assert. Hence we remove the duration argument.

Since the modem can start sending traces before the
nrf_modem_lib_start() is called, any maximum size configured by the
application may be ignored when the first few traces are received. This
leads to erronueous calculation of number of trace bytes received. It is
better for the application to handle the size of trace data received and
stop using the nrf_modem_lib_trace_stop() API when needed.

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>