EMDS: Implement emergency data storage
Create the emergency data storage. This storage should be have
its own dedicated storage area that can be quickly stored to at when
necessary. This way we can store data that is changing a lot without
wearing out the flash, and have a predefined time for storing the
data. There should be possible to determine how mush time is needed
to store all data configured to be stored in this area.

Signed-off-by: Ingar Kulbrandstad <ingar.kulbrandstad@nordicsemi.no>